requirements.txt
streamlit>=1.30.0
slack_sdk>=3.27.0
openai>=1.0.0
python-dotenv>=1.0.0
supabase>=2.0.0
